Refactor teacher dashboard into modular components
